fix: histogram plotter bugs - hist: wrong behaviour after change in settings - hist: handle negative input in settings - triState: argument type chenge to save memory - texture plotter: spelling

+++ Interface/ImguiExtensions.cpp
@@ -278,7 +278,7 @@ namespace ImGui {
         }
 
     }
-    bool TriState(const char* label, size_t* v, bool allowManualMixed) {
+    bool TriState(const char* label, short* v, bool allowManualMixed) {
         ImGuiWindow* window = GetCurrentWindow();
         if (window->SkipItems)
             return false;
